refactor text inputs
